+++ README.md
@@ -1,4 +1,6 @@
-# discord-some-cache
+# discord-no-cache
 Alternative WebSocket and small restful for using the discord client so there's no caching involved (internally).
+
+
 
 For internal use of simplistic libs that don't need caching at all.
