+++ index.js
@@ -1,0 +1,3 @@
+const Client = require("./Client.js");
+
+module.exports = Client;
