--- src/index.js
@@ -1,7 +1,0 @@
-const Client = require("./lib/Client");
-const Channel = require("./lib/Channel");
-const Embed = require("./lib/Embed");
-const Message = require("./lib/Message");
-const Shard = require("./lib/Shard");
-
-module.exports = { Client, Channel, Embed, Message, Shard };
