--- src/utils/wait.js
@@ -1,1 +1,0 @@
-module.exports = function wait(a) { return new Promise(r => { setTimeout(() => r(), a); }); };
